LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/window-open-noopener__top-expected.txt

CONSOLE MESSAGE: TypeError: this.checkValues is not a function. (In 'this.checkValues(w2)', 'this.checkValues' is undefined)

Harness Error (TIMEOUT), message = null

NOTRUN noopener window.open targeting _top
